requirements.txt
numpy==1.19.3
Flask==1.1.2
gunicorn==20.0.4
Pillow==8.0.1
keras==2.4.3
tensorflow==2.4.1
h5py==2.10.0
